LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/referrer-policy-no-referrer.h2.window-expected.txt

CONSOLE MESSAGE: <link rel=preload> has an invalid `href` value

FAIL Referrer policy: no-referrer assert_true: expected true got false
